+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+- Native support for deploying permanent Databricks Views with input data quality evaluation and Dataset Resolution using `declare_with_governance` within the Spark integration helpers.
+- Added support for `is_null` data quality validation rule to natively assert column emptiness avoiding SQL Tri-Valued logic evaluation issues.
+
 ## [0.41.0.0] - 2026-03-31
 
 ## [0.40.0.0] - 2026-03-19